Rename `CSVExportWidgetSelection` to `ExportWidgetSelection`.
