+++ graylog2-web-interface/src/views/components/export/ExportWidgetSelection.tsx
